requirements.txt
Flask==0.12.2
Flask-SSLify==0.1.5
gunicorn==19.7.1
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-cloud-storage==1.10.0
requests==2.19.1
